dtnd configuration.toml: Working example
As addressed in #29, the default configuration enables lots of features
which are neither required nor available in most environments. This
commit changes this behaviour by reducing the configuration to a subset
which should work on most systems. The more specific sections are
commented out, which should be the way to go for further modifications.